Fix handling for /*/contributors.txt routes (#2287)
This change fixes a problem that causes Yari to crash for requests
to /*/contributors.txt routes. The existing code expects a call to
`await buildDocument(document)` to return an array. when in fact it
returns an object.

Without this change, requests for /*/contributors.txt routes cause Yari
to crash with this log output:

node_modules/@mdn/yari/server/index.js:102
  const [builtDocument] = await buildDocument(document);
                          ^
TypeError: (intermediate value) is not iterable
    at node_modules/@mdn/yari/server/index.js:102:27
error Command failed with exit code 1.